JS/TS: export-usage binder via ImportInfo
